splitting files further into annotations/aberrations so that common ops can be used on votes
