+++ routes/sampleView.js
@@ -3,7 +3,8 @@
 var mongoose = require( 'mongoose' ),
 	Database = require('../model/db'),
 	Samples = require('../model/samples'),
-        Annotations = require('../model/annotations'),
+        Base_annotations = require('../model/annotations'),
+Aberrations = require('../model/aberrations'),
 	Cancers = require('../model/cancers'),
 	Datasets = require('../model/datasets'),
 	fs = require('fs');
@@ -61,7 +62,7 @@ exports.sampleView = function sampleView(req, res){
 			    geneMutations = [];
 
 			    // call for additional mutations
-			    Annotations.geneFind(Annotations.inGeneClause('gene', mutGenes),'right', function(err, userAnnos) {
+			    Base_annotations.geneFind(Base_annotations.inGeneClause('gene', mutGenes),'right', function(err, userAnnos) {
 				if (err) {
 				    console.error(err);
 				    fail = true;
